[APL] Make EPAYLOAD size dynamic
On APL platform, default EPAYLOAD size is more than 1MB to enable
UEFI payload. However, if UEFI payload is not used, the space is
still reserved in flash. A more flexible way is to reserve the
EPAYLOAD size only when it is specified in the build command line.
This patch added this capability.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>