[APL] Adjust flash layout padding for stitching
SBL supports open source stitching mechanism based on an existing
IFWI binary. It can also support FIT stitching mechanism. For FIT
stitching, it is required to add certain padding in the layout so
that each component in IBB and OBB partition can be 4KB aligned.
Current padding component IPAD/OPAD size needs to be updated to
match the number of the component inside IBB and OBB partition.
This patch adjusted the padding size accordingly.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>