Add duplicated CFGDATA check in GenCfgData.py
CFGDATA name should be unique in the whole CFGDATA database. Current
tool does not check this and will keep silent for this error. This
brings issue into later development stage. The build process should
be enhanced to check this and error out if duplicated CFGDATA item is
found. This patch added this check.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>