[APL] Update StitchLoader.py to check Boot Guard status
The APL StitchLoader.py script has its own limitation. For example,
if Boot Guard is enabled in the base image, the final stitched image
with this script might not boot due to security reason. It is better
to check this during stitching process so that it can error out earlier
to notify the user.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>