[APL] Fix ConfigEditor error on changing Boot Option
To ensure ConfigEditor can handle the configuration data type properly,
the default value needs to be valid. For string type configuration item,
the value should start with single quote. The APL default values for OS
image name is set to number 0 instead of '' to indicate empty string.
This patch corrected this and fixed #151.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>