[CFL] Optimize Gpio Cfg Data input
PinOffset information like PadNum, GrpIdx for a GPIO can be
adjusted in the fields available in DW1, where as ChipsetId
can be determined dynamically. This reduces the config data
size significantly. So, we can remove PinOffset info i.e. the
first dword in the GpioTableData.

Moved static definitions to top of the file in Stage2 board init.

Signed-off-by: Sai Talamudupula <sai.kiran.talamudupula@intel.com>