[CFL] MRC data save is ignored after first boot
Current code will check if the first 4 bytes of the existing
MRC training data, if it is 0xFFFFFFFF then MRC training data
will be saved otherwise even though there is change in the
training data, it will not be saved.

Modified the code to compare the incoming training data to the
saved data, if they do not match, save the incoming training data.

Signed-off-by: Raghava Gudla <raghava.gudla@intel.com>