[CFL] Add Native Function help in Gpio config (#316)
Current Gpio config provides a way to select PadMode. But
Native Functions for a PCH are just numbered #1,2, etc. and
do not provide enough information for the end user while
configuring the PadMode.

This patch adds the required Native Function information help
strings in the drop down menu while configuring the PadMode.

Signed-off-by: Sai Talamudupula <sai.kiran.talamudupula@intel.com>