Add GenGpioData.py tool
GenGpioData.py is a utility that converts
GPIO pin data from one format to other.

The formats currently supported are [h, csv, txt, dsc, dlt].
[h, csv, txt] formats are external to SBL and [dsc, dlt] formats
are known to SBL. So, this tool provides a way to convert one
of the [h, csv, txt] to [dsc, dlt] and vice-versa.

Example usage:

    python GenGpioData.py -if GpioInput.csv -of dsc

    python GenGpioData.py -if GpioInput.dsc -of txt

Signed-off-by: Sai Talamudupula <sai.kiran.talamudupula@intel.com>