Use Treeview widget to display table in ConfigEditor
Current ConfigEditor uses multiple Entry widgets to build a Table
widget since Tkinter does not support Table widget. However, when
the table is big, the loading performance is very slow because of
large mount of Entry widgets. This patch switched to use Treeview
widget to build a table widget, and it is much more faster than the
original implementation.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>