[WHL] Allow relative path in StitchIfwi.py
Currently, only absolute path is allowed in '-w' stitch_dir and '-s'
stitch_zip file. In addition, if stitch_dir is not absolute path, the
stitch tool looks for 'SBL_SOURCE' which does not exist in OS environment.

This patch allows the StitchIfwi.py tool to use relative path as well.

Change-Id: I08826406995a8d81a01be6951169b1302e5aafb7
Signed-off-by: Aiden Park <aiden.park@intel.com>