Fix OsLoader grub config parsing issue (#607)
This is a regression on the grub.cfg file parsing. When parsing
each menu entry in grub.cfg, the index should not be advanced
at the end. It will cause the boot index shift by 1 if doing so.
This patch fixed this issue. Tested with Ubuntu boot.
It fixed issue #606.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>