Call DeInit() for every boot device (#658)
Current SBL OsLoader call DeinitBootDevices() once at the end of
the boot option loop. This is not right because it will only DeInit
the current active boot device. Instead, DeInit() should be called
inside the loop for every boot device. Specially, if USB is used as
console, USB DeInit() cannot be called at the moment. It will be
deferred until OS transfering.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>