Add support to report free memory resource (#665)
This patch added support to report free memory resource lenghth.
It will search for all used memory pages and add them together.
The "virtual" free address will be returned to indicate the
virtual start point of the free memory top. It is virtual since the
memory allocation can be fragmented. This is just an indicator to
calculate the actual used memory size:
  UsedMemSize = EndAddr - FreeAddr

Signed-off-by: Maurice Ma <maurice.ma@intel.com>