Fix an issue in firmware update
This patch fixed an issue in firmware update. During the saved capsule
signature verification against the capsule signature to make sure
that the capsule did not change during the reboot, instead of using
length of the signature, used the size of the macro that indicate the
length of the signature.

Verified that firmware update is able to pass on WHL.

Signed-off-by: Raghava Gudla <raghava.gudla@intel.com>