Resolve raw partition boot loading
The calculation on where to start reading the
remainder of the boot image from is not being
calculated properly, the aligned header size
is given in bytes instead of in terms of block
size. This patch updates the calculation
accordingly.

Also, if SwPart is set as 255 then the LBA given
will be used as an absolute LBA instead of used
relative to the SwPart's starting LBA.

Signed-off-by: James Gutbub <james.gutbub@intel.com>