Check for '.exe' in run_process routine
run_process is used in stitching scripts
and other places and when the command being
called might also have a similarly named
command in the same folder (ex. fit and fit.exe)
where one command is meant to be ran in Linux
and the other is meant to be ran in Windows, there
can be issues that require to check for which
OS the stitching script is running in to append
'.exe' if Windows is detected. Rather than change
it every where else we can change it in the
run_process routine to check for Windows and if
there is not already and extestion included in the
command we can append '.exe'.

Also need to resolve an issue with the OpenSSL path
during GenerateKeys; since the path in Azure is in
Program Files instead of in C:\OpenSSL it is adding
'.exe' to the wrong location in the filepath.

Signed-off-by: James Gutbub <james.gutbub@intel.com>