[APL] Enable IA untrusted mode at end of SBL stage
Current APL SBL code will enable IA_UNTRUSTED mode only at end of
firmware notification. It might be too late for certain conditions.
This patch moves it to be set at end of stage in SBL. In this way,
it ensures the bit is set before launching any external payload.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>