Fix firmware update version check
As part of the commit e21c74f42ef7c455f22f34c9f2137e7747b0d3aa,
it used the HOB to get the firmware version info. However, the
HOB firmware version structure is not exactly the same as SBL
firmware version structure. It needs to be adjusted accordingly.
This patch addressed this issue.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>