[WHL] Add mailbox debug port example code
This patch provides exampole on how to enable mailbox debug port on
WHL. To enable it, the following needs to be set in BoardConfig.py.
  self.DEBUG_OUTPUT_DEVICE_MASK = 0x07
  self.CONSOLE_IN_DEVICE_MASK   = 0x00000005
  self.CONSOLE_OUT_DEVICE_MASK  = 0x00000005
Also adjust the PcdMailboxDebugPortMmioBase value in BoardConfig.py
as needed.

On host side, python script can be used in ISD to read/write this
register using IPC interfaces. EX:
  import ipccli
  ipc = ipccli.baseaccess()
  ipc.stateport.cnp_tpsb0.tap2iosfsb32 (...)
Please refere to MailboxDebugPort.py for example.

This was tested on UPX board with CCA debuger connected.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>