Adjust PCI resource check condition
Current PciBus lib expects the root bridge resource base should not be
greater than limit. It is true for normal case. However, to mark the
source is unavailable, the base could be less than the limit in some
case. PCI bus lib should only validate the resource that does have a
request.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>