Add support to show a configure item under multiple pages
Current CFGDATA item can only be displayed under one single page
defined in YAML. In some special case it might need to duplicate
the same item under a differnt page. This patch added this feature
by introducing a new "altpage" attribute for a CFG item. When a
item needs to be displayed on a 2nd page, 'altpage' can be added
to indicate the 2nd page id.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>