Enhance memory gap handling in memory map
The UEFI payload uses the memory map info to guess the top of low
memory. It expects the memory map to be continuous. However, in
certain case, the memory resource HOB produced by FSP is not
continous, and it causes issue for UEFI payload to find the correct
top of low memory. This patch enhanced SBL memory map handling to
check if there is any gap in the memory map. If so, it will try to
insert a new entry to fill the gap.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>