Fix GPIO CFGDATA database overriding issue
Current GPIO programming code in CommonSocPkg will override the
CFGDATA database. It tried to zero out the GPIO group ID and pad
ID from the GPIO CFGDATA entry. But it should only be applied to
the copied new entry, not the original entry in CFGDATA database.
This patch fixed this isue.
It fixed #1382.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>