[TGL] Fix the GPIO group ID overriding issue
Current SBL uses DLT file to override the GPIO group id in the
GPIO CFGDATA table because the group ID used in CFGDATA needs to
match the group ID used by GPIO library. This patch decoupled the
GPIO group id with the GPIO library. Instead, a translation was
added to convert the group ID to the value required by the GPIO
library.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>