Add new Constant type for CFGDATA
There are CFGDATA item that needs to be generated into DLT file,
but not to be displayed in ConfigEditor. The current "Reserved"
type prevents the item from being displayed in ConfigEditor,
but also prevents it from being generated in a DLT file. The new
"Constant" type is introduced here to resolve this issue. Item
with "Constant" type will not be displayed in ConfigEditor, but
will be generated in the DLT file.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>